fix(KEN-1130): Address PR review - generate the bundle grammar
The two key-coverage guards did not hold: both searched a text whose
opening sentence already lists the item folders, so every key and every
singularization of one was already a substring of something unrelated.
Renaming a MEMBER_LISTS key left both green.

Closed structurally instead of with a third matcher. The [bundles] sentence
in the kendex init marker and in the marketplace new scaffold is built from
MEMBER_LISTS, and so is the scaffold's example, which now carries one member
of every kind. Both substring guards are gone; what remains is the round
trip, which reddens when the reader stops seeing a key the scaffold writes,
and the golden digests, which redden on the key rename. Digests refreshed.

The own_catalog module doc blamed request.no_auto_skills for the missing
agent-to-skill expansion. That flag defaults false and only the desktop app
sets it; the reason that holds on both paths is that the expansion walks
request.agents, which a set's members never join. The A_MAPPING comment
credited a prefix rule that yields "arch" and finds nothing; role-skills is
what reaches reviewer.

workflow is the three sets plus deep-research, not those three sets. Said
that way in kendex.toml, the changelog and the doc comments, and the
containment test now runs both ways against a written-down extras list, so
an undescribed member reddens.

The agent walk asserts every name [agent-skills] and [role-skills] give a
member is a skill the catalog offers. upstream_skills drops a name it cannot
resolve, so a typo there was silent.

own_catalog moved to its own file: the assertions took bundles.rs past the
size ratchet.
