+++ changelog.d/added/KEN-1130.md
@@ -1,1 +1,1 @@
-- New `workflow` bundle: orchestration, code-review and commit-guards in one `kendex add --bundle workflow`, 30 agents, skills and hooks that carry what each other needs.
+- New `workflow` bundle: orchestration, code-review and commit-guards plus `deep-research` in one `kendex add --bundle workflow`, 30 members that carry what each other needs.
