+++ crates/core/src/author/scaffold.rs
@@ -225,15 +225,19 @@ fn manifest_text(name: &str, description: &str, author: &str, license: License) 
     if let Some(spdx) = license.spdx() {
         text.push_str(&format!("license = {}\n", toml_string(spdx)));
     }
-    text.push_str(
+    // The `[bundles]` keys come from the list that reads them: a catalog
+    // author following a hand-written copy is how the shape shipped wrong.
+    text.push_str(&format!(
         "\n# Items live in agents/, skills/, hooks/, commands/ and mcp/.\n\
          # Curated sets: [bundles.<name>] with a description and one list of\n\
-         # bare names per kind — agents, skills, commands, hooks, mcp-servers.\n\
+         # bare names per kind: {}.\n\
          # [bundles.starter]\n\
-         # description = \"Everything a new project needs\"\n\
-         # skills = [\"my-skill\"]\n\
-         # agents = [\"my-agent\"]\n",
-    );
+         # description = \"Everything a new project needs\"\n",
+        crate::source::bundles::member_list_keys()
+    ));
+    for line in crate::source::bundles::member_list_example().lines() {
+        text.push_str(&format!("# {line}\n"));
+    }
     text
 }
